refactor: move bots logic to src
